requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
django-filter==0.13.0
django-material==0.9.1
djangorestframework==3.4.7
enum34==1.1.6
gunicorn==19.6.0
heroku==0.1.4
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
MySQL-python==1.2.5
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.9.0
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==1.5
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
traitlets==4.3.0
wcwidth==0.1.7
